Configure Terraform provider and import resource
Configure Terraform provider (profile, region, etc.) is required to use import and
later other functionality from the provider.

The import API requires as input the ID to uniquely identify a resource
and returns a state for the resource that is mostly empty. Read needs to
be called afterwards to fill all the state attributes, which are retrieved
 from the real resource in AWS.