Hide Terraform state and provider internals
Main function uses now high-level functions only without needing to
 know about Terraform state or provider internal structures and functions.
Internals are hidden behind API in state.go and provider.go.

For state.go, for example, high-level methods exposed are a function to get
all provider names in the state (eg. aws, google) as well as a function returning all
resources of a state in a custom structure with only the attributes
needed for deleting the resources (Type, ID, Mode, ProviderName).