Skip calling Hashicorp server when provider binary exists
This is useful to avoid a costly NAT gateway when running in
an AWS lambda and the provider binary has been copied already
to the right local path beforehand (kind of an offline mode).

Note: caching of provider binaries wasn't sufficient, because
an internet connection (to outside of AWS) was still
needed to call the Hashicorp server for checksums, etc.